Adjustments, Additions, and Fixes
Toyed around with adding a button to the embed message (in the future, this button will allow the user to toggle between a path that includes getting a ring, and one without).

Removed old features/commented out bits and pieces of code that was made for the Easter event back in April, for clarity sake (they were taking up a lot of space).

Configured the bot to reply with the reason why it could not parse the map (i.e. if something went wrong searching for the Juice resource, it will say so in the embed message).

Made it so the embed message footer says what user the solution is for.