Solution Message Change
To make things easier for the user, the solution message will now have a button on it, that when pressed, brings up alternate paths, instead of posting all paths (There was a problem where a user mixed up both paths, and ended up failing the overall date game).

The message's buttons will also be on a 3-minute timer; once the message is 3 minutes old, the message is considered 'timed out', and will not be serviced.

Currently, if there is no ring, the airport solution will also be included. However, if there is a ring, we assume that the user is satisfied, and will not want to reach the airport to 'reroll' the board.